fix(conv): gate the 1x1-fast dsp leg and fold the 1xN routing arithmetic to 64 bits (#371)
Two proven conv-sizer defects from #364's review, fixed at their roots.

The 1x1-fast _dsp helper's gate lived only in its armclang branch, so on
every other compiler the leg returned 0 reading nothing -- the wrapper
answered 0 where its dispatcher answered -1 (issue: the divergence class
of the sizer sweep, on a symbol that sweep did not name; the Python
bindings call these legs on exactly such builds). The gate now sits
ahead of the compiler split, and the generic 1x1-fast sizer validates
the same bound once ahead of its dispatch: dispatcher, _dsp and _mve
answer out-of-range 1x1-fast shapes identically on every target and
compiler, and the README leg-contract sentence promises exactly that --
-1 wherever the dispatcher diagnoses one -- with the old exception
clause gone.

arm_nn_is_convolve_1_x_n multiplied stride.w * c in int32 -- signed
overflow UB, UBSan-proven reachable from public sizers at 65536 x 65536.
Folded to int64, answer unchanged for every non-overflowing input
(independent sweeps: 20M and 21,600 probes, 0 mismatches). Review found
the same multiply open-coded in four runtime sites; all folded, and the
s8 wrapper's 1xN route now calls the predicate exactly like its s4
sibling, so sizer and runtime routing agree by construction.

Two Unity cases pin both fixes (red against the old code; the 1x1
case's cross-leg equality is asserted armclang-safely -- the _dsp byte
count is compiler-conditional, non-negativity is the cross-build
invariant).

Refs #366, #367 -- both fully resolved; closed by hand with this merge.
Reviewed: two lenses + independent fix verification (21,600-probe sweep)
+ final-gate review with its own ground-truth probes; logs on PR #371.

+++ Source/ConvolutionFunctions/arm_convolve_1_x_n_s4.c
@@ -62,7 +62,7 @@ arm_cmsis_nn_status arm_convolve_1_x_n_s4(const cmsis_nn_context *ctx,
     int32_t buffer_size = arm_convolve_1_x_n_s4_get_buffer_size(conv_params, input_dims, filter_dims, output_dims);
     /* The wrapper API is the ultimate reference for argument check */
     if ((input_dims->h != 1) || conv_params->dilation.w != 1 || (buffer_size != 0 && ctx->buf == NULL) ||
-        conv_params->stride.w == 0 || (conv_params->stride.w * input_dims->c % 4 != 0))
+        conv_params->stride.w == 0 || (((int64_t)conv_params->stride.w * input_dims->c) % 4 != 0))
     {
         return ARM_CMSIS_NN_ARG_ERROR;
     }
